Enforce Security by removing the Package Manager
Hey there,

We want to suggest to remove the Package Manager after installing and compiling everything needed. This would remove the capability in case there is a security problem in the future to give nobudy more tools then needed. This would also mean that future apk related security holes would not affect this project. We did some tests with our repo (and off course live testing in our enviroment) where we added this in the Dockerfile and so far it has none negative affect to the build or running the application :D . It could also be possible to include the change into an existing RUN command to avoid creating a new layer in the image.

Here you can see our current build, to check out what we try out:
https://github.com/kgncloud/mautrix-whatsapp

Best Regards,

KN Team